+++ presentation/src/main/java/com/strone/presentation/model/OrderbookModel.kt
@@ -4,13 +4,14 @@ import android.os.Parcelable
 import androidx.compose.runtime.Immutable
 import com.strone.domain.model.type.OrderType
 import kotlinx.parcelize.Parcelize
+import java.math.BigDecimal
 
 @Parcelize
 @Immutable
 class OrderbookModel(
     val code: String, // 마켓 코드 ex) KRW-BTC
-    val totalAskSize: Double, // 호가 매도 총 잔량
-    val totalBidSize: Double, // 호가 매수 총 잔량
+    val totalAskSize: BigDecimal, // 호가 매도 총 잔량
+    val totalBidSize: BigDecimal, // 호가 매수 총 잔량
     val orderbookUnitModels: List<OrderbookUnitModel>, // 호가 정보
     val timestamp: Long, // 타임스탬프
 ): Parcelable {
@@ -18,8 +19,8 @@ class OrderbookModel(
     @Parcelize
     @Immutable
     data class OrderbookUnitModel(
-        val price: Double,
-        val size: Double,
+        val price: BigDecimal,
+        val size: BigDecimal,
         val orderType: OrderType
     ): Parcelable
 }
